+++ src/TRegx/CleanRegex/Replace/ReplacePattern.php
@@ -2,13 +2,12 @@
 namespace TRegx\CleanRegex\Replace;
 
 use TRegx\CleanRegex\Exception\NotReplacedException;
-use TRegx\CleanRegex\Match\FindFirst\Optional;
 
-interface ReplacePattern extends SpecificReplacePattern, Optional
+interface ReplacePattern extends SpecificReplacePattern
 {
-    public function orThrow(string $exceptionClassName = NotReplacedException::class): SpecificReplacePattern;
+    public function throwingOtherwise(string $exceptionClassName = NotReplacedException::class): SpecificReplacePattern;
 
-    public function orReturn($substitute): SpecificReplacePattern;
+    public function returningOtherwise($substitute): SpecificReplacePattern;
 
-    public function orElse(callable $substituteProducer): SpecificReplacePattern;
+    public function otherwise(callable $substituteProducer): SpecificReplacePattern;
 }